requirement.txt
newsapi-python==0.2.7
pyTelegramBotAPI==4.14.0
transformers==4.34.1
newspaper3k==0.2.8
beautifulsoup4==4.12.2
requests==2.31.0
torchaudio==2.1.0
torchvision==0.16.0
torch==2.1.0